+++ recipes/adoptopenjdk.rb
@@ -3,7 +3,7 @@ Chef::Log.fatal("
 
 Adopt OpenJDK recipe is now deprecated
 Using the adoptopenjdk_install resource is now recommended
-See: documentation/resources/install_adoptopenjdk.md for help
+See: https://github.com/sous-chefs/java/blob/master/documentation/resources/adoptopenjdk_.md for help
 
 ")