+++ resources/corretto_install.rb
@@ -1,16 +1,15 @@
-resource_name :adoptopenjdk_install
-include Java::Cookbook::AdoptOpenJdkHelpers
-default_action :install
+resource_name :corretto_install
+include Java::Cookbook::CorrettoHelpers
 
 property :version, String, name_property: true,
                            description: 'Java version to install'
-property :variant, String, equal_to: %w(hotspot openj9 openj9-large-heap), default: 'openj9',
-                           description: 'Install flavour'
-property :url, String, default: lazy { default_adopt_openjdk_url(version)[variant] },
+property :full_version, String,
+  description: 'Used to configure the package directory, change this is the version installed by the package is no longer correct'
+property :url, String, default: lazy { default_corretto_url(version) },
                        description: 'The URL to download from'
-property :checksum, String, regex: /^[0-9a-f]{32}$|^[a-zA-Z0-9]{40,64}$/, default: lazy { default_adopt_openjdk_checksum(version)[variant] },
+property :checksum, String, regex: /^[0-9a-f]{32}$|^[a-zA-Z0-9]{40,64}$/, default: lazy { default_corretto_checksum(version) },
                             description: 'The checksum for the downloaded file'
-property :java_home, String, default: lazy { "/usr/lib/jvm/java-#{version}-adoptopenjdk-#{variant}/#{sub_dir(url)}" },
+property :java_home, String, default: lazy { "/usr/lib/jvm/java-#{version}-corretto/#{corretto_sub_dir(version, full_version)}" },
                              description: 'Set to override the java_home'
 property :java_home_mode, String, default: '0755',
                                   description: 'The permission for the Java home directory'
@@ -20,7 +19,7 @@ property :java_home_group, String, default: lazy { node['root_group'] },
                                    description: 'Group for the Java Home'
 property :default, [true, false], default: true,
                                   description: ' Whether to set this as the defalut Java'
-property :bin_cmds, Array, default: lazy { default_adopt_openjdk_bin_cmds(version)[variant] },
+property :bin_cmds, Array, default: lazy { default_corretto_bin_cmds(version) },
                            description: 'A list of bin_cmds based on the version and variant'
 property :alternatives_priority, Integer, default: 1,
                                           description: 'Alternatives priority to set for this Java'
@@ -53,7 +52,8 @@ action :install do
 
   node.default['java']['java_home'] = new_resource.java_home
 
-  template "/usr/lib/jvm/.java-#{new_resource.version}-adoptopenjdk-#{new_resource.variant}.jinfo" do
+  # Set up .jinfo file for update-java-alternatives
+  template "/usr/lib/jvm/.java-#{new_resource.version}-corretto.jinfo" do
     cookbook 'java'
     source 'jinfo.erb'
     owner new_resource.java_home_owner
@@ -87,7 +87,7 @@ action :remove do
     action :unset
   end
 
-  directory "AdoptOpenJDK removal of #{extract_dir}" do
+  directory "Removing #{extract_dir}" do
     path extract_dir
     recursive true
     only_if { ::File.exist?(extract_dir) }