+++ test/fixtures/cookbooks/test/recipes/corretto.rb
@@ -1,0 +1,1 @@
+corretto_install node['version']