+++ test/fixtures/cookbooks/test/recipes/openjdk.rb
@@ -1,0 +1,1 @@
+openjdk_install node['version']