+++ test/fixtures/cookbooks/test/recipes/openjdk_direct.rb
@@ -1,3 +1,0 @@
-include_recipe 'test::base'
-include_recipe 'java::openjdk'
-include_recipe 'test::java_cert'